Re #2407 move ProjectPackage helpers out of Stack.Types.Config
